requirements.txt
discord.py
tensorflow
transformers
python-dotenv
pandas
scikit-learn